Preferences: check ML state for clearDatabase
Instead of stopping MediaParsingService before clearing the
medialibrary to ensure that there won't be any operations going, this
just checks if the medialibrary is working before clearing the database

+++ application/television/src/main/java/org/videolan/television/ui/preferences/PreferencesAdvanced.kt
@@ -136,29 +136,47 @@ class PreferencesAdvanced : BasePreferenceFragment(), SharedPreferences.OnShared
                 return true
             }
             "clear_media_db" -> {
-                val dialog = ConfirmDeleteDialog.newInstance(title = getString(R.string.clear_media_db), description = getString(R.string.clear_media_db_message), buttonText = getString(R.string.clear))
-                dialog.show((activity as FragmentActivity).supportFragmentManager, RenameDialog::class.simpleName)
-                dialog.setListener {
-                    launch {
-                        val medialibrary = Medialibrary.getInstance()
-                        activity.stopService(Intent(activity, MediaParsingService::class.java))
-                        withContext((Dispatchers.IO)) {
-                            medialibrary.clearDatabase(false)
-                            deleteAllWatchNext(activity)
-                            //delete thumbnails
-                            try {
-                                activity.getExternalFilesDir(null)?. let {
-                                    val files = File(it.absolutePath + Medialibrary.MEDIALIB_FOLDER_NAME).listFiles()
-                                    files?.forEach { file ->
-                                        if (file.isFile) FileUtils.deleteFile(file)
+                val medialibrary = Medialibrary.getInstance()
+                if (medialibrary.isWorking) {
+                    activity?.let {
+                        Toast.makeText(
+                            it,
+                            R.string.settings_ml_block_scan,
+                            Toast.LENGTH_LONG
+                        ).show()
+                    }
+                } else {
+                    val dialog = ConfirmDeleteDialog.newInstance(
+                        title = getString(R.string.clear_media_db),
+                        description = getString(R.string.clear_media_db_message),
+                        buttonText = getString(R.string.clear)
+                    )
+                    dialog.show(
+                        (activity as FragmentActivity).supportFragmentManager,
+                        RenameDialog::class.simpleName
+                    )
+                    dialog.setListener {
+                        launch {
+                            activity.stopService(Intent(activity, MediaParsingService::class.java))
+                            withContext((Dispatchers.IO)) {
+                                medialibrary.clearDatabase(false)
+                                deleteAllWatchNext(activity)
+                                //delete thumbnails
+                                try {
+                                    activity.getExternalFilesDir(null)?.let {
+                                        val files =
+                                            File(it.absolutePath + Medialibrary.MEDIALIB_FOLDER_NAME).listFiles()
+                                        files?.forEach { file ->
+                                            if (file.isFile) FileUtils.deleteFile(file)
+                                        }
                                     }
+                                    BitmapCache.clear()
+                                } catch (e: IOException) {
+                                    Log.e(this::class.java.simpleName, e.message, e)
                                 }
-                                BitmapCache.clear()
-                            } catch (e: IOException) {
-                                Log.e(this::class.java.simpleName, e.message, e)
                             }
+                            medialibrary.discover(AndroidDevices.EXTERNAL_PUBLIC_DIRECTORY)
                         }
-                        medialibrary.discover(AndroidDevices.EXTERNAL_PUBLIC_DIRECTORY)
                     }
                 }
                 return true
