+++ application/vlc-android/src/org/videolan/vlc/gui/preferences/PreferencesAdvanced.kt
@@ -42,15 +42,11 @@ import androidx.preference.EditTextPreference
 import androidx.preference.Preference
 import kotlinx.coroutines.*
 import org.videolan.medialibrary.interfaces.Medialibrary
-import org.videolan.resources.AndroidDevices
-import org.videolan.resources.KEY_AUDIO_LAST_PLAYLIST
-import org.videolan.resources.KEY_MEDIA_LAST_PLAYLIST
-import org.videolan.resources.VLCInstance
+import org.videolan.resources.*
 import org.videolan.tools.BitmapCache
 import org.videolan.tools.Settings
 import org.videolan.tools.putSingle
 import org.videolan.vlc.BuildConfig
-import org.videolan.vlc.MediaParsingService
 import org.videolan.vlc.R
 import org.videolan.vlc.gui.DebugLogActivity
 import org.videolan.vlc.gui.dialogs.ConfirmDeleteDialog
@@ -118,31 +114,48 @@ class PreferencesAdvanced : BasePreferenceFragment(), SharedPreferences.OnShared
                 return true
             }
             "clear_media_db" -> {
-                val dialog = ConfirmDeleteDialog.newInstance(title = getString(R.string.clear_media_db), description = getString(R.string.clear_media_db_message), buttonText = getString(R.string.clear))
-                dialog.show(requireActivity().supportFragmentManager, RenameDialog::class.simpleName)
-                dialog.setListener {
-                    lifecycleScope.launch {
-                        val medialibrary = Medialibrary.getInstance()
-                        requireActivity().stopService(Intent(requireActivity(), MediaParsingService::class.java))
-                        withContext((Dispatchers.IO)) {
-                            medialibrary.clearDatabase(false)
-                            //delete thumbnails
-                            try {
-                                requireActivity().getExternalFilesDir(null)?. let {
-                                    val files = File(it.absolutePath + Medialibrary.MEDIALIB_FOLDER_NAME).listFiles()
-                                    files?.forEach { file ->
-                                        if (file.isFile) FileUtils.deleteFile(file)
-                                    }
-                                }
-                                BitmapCache.clear()
-                            } catch (e: IOException) {
-                                Log.e(this::class.java.simpleName, e.message, e)
-                            }
-                        }
-                        medialibrary.discover(AndroidDevices.EXTERNAL_PUBLIC_DIRECTORY)
+                val medialibrary = Medialibrary.getInstance()
+                if (medialibrary.isWorking) {
+                    activity?.let {
+                        Toast.makeText(
+                            it,
+                            R.string.settings_ml_block_scan,
+                            Toast.LENGTH_LONG
+                        ).show()
                     }
+                } else {
+                    val dialog = ConfirmDeleteDialog.newInstance(
+                        title = getString(R.string.clear_media_db),
+                        description = getString(R.string.clear_media_db_message),
+                        buttonText = getString(R.string.clear)
+                    )
+                    dialog.show(
+                        requireActivity().supportFragmentManager,
+                        RenameDialog::class.simpleName
+                    )
+                    dialog.setListener {
+                        lifecycleScope.launch {
+                            withContext((Dispatchers.IO)) {
+                                medialibrary.clearDatabase(false)
+                                //delete thumbnails
+                                try {
+                                    requireActivity().getExternalFilesDir(null)?.let {
+                                        val files =
+                                            File(it.absolutePath + Medialibrary.MEDIALIB_FOLDER_NAME).listFiles()
+                                        files?.forEach { file ->
+                                            if (file.isFile) FileUtils.deleteFile(file)
+                                        }
+                                    }
+                                    BitmapCache.clear()
+                                } catch (e: IOException) {
+                                    Log.e(this::class.java.simpleName, e.message, e)
+                                }
+                            }
+                            medialibrary.discover(AndroidDevices.EXTERNAL_PUBLIC_DIRECTORY)
+                        }
+                    }
+                    return true
                 }
-                return true
             }
             "clear_app_data" -> {
                 if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
